requirements.txt
# Kaggle Benchmarks SDK
kaggle-benchmarks>=0.1.0

# Kaggle API
kaggle>=1.5.16

# Utilidades
pandas>=2.0.0
numpy>=1.24.0
jsonschema>=4.17.0
tqdm>=4.65.0

# Testing
pytest>=7.3.0

# Para generación de datos
faker>=18.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0